Fix for issue #1778.  The ITextView which is created for the WpfConsole previously didn't explicitly specify an ITextViewRole[] argument.  This means it inherits the default values which is that of a document.  In particular it would get the roles Document and Structured.  Because the IContentType of the ITextBuffer was derived from text it caused all extensions to see it simply as a text document.
This creates a problem because it's not an editable document.  This causes extensions / features to incorrectly target the window.  For example Visual Studio will try to apply outlining and extensions like VsVim will appear in the window.

Instead it should offer roles similar to that of the immediate window, the command window and interactive window.  This will allow for extensions to correctly target it