Support pre-installed packages where the project template loads the package source location from the Registry.
The <packages> node now supports a repository value of "registry" and that
then requires a new "keyName" attribute.  This key name indicates the registry
key to get the packages path value from, within the following location:

HKCU\SOFTWARE\NuGet\Repository\[keyName], with fall-back to:
HKLM\SOFTWARE\NuGet\Repository\[keyName]

This allows installers to simply lay nupkg files down on disk, and register
a key that points to that location, rather than having to utilize a VSIX that
contains the nupkgs.  With this approach, satellite packages can be installed
within the same repository as the runtime packages.