Fix for issue 2125
The GetPackagesPath function was incorrectly assuming that there would
be a repositoryPath element in the NuGet.config file.  This is not
necessarily true as it's a user controlled XML file.  It is also not
added in core scenarios where the user has only run a limited set of
NuGet.exe commands.  If the element wasn't there NuGet.exe would
terminate with a NullReferenceException

This fix addresse the problem by properly checking the XmlElement for
null.