Move from a byte[] based hash calculation to the stream methods. In most places we already have a MemoryStream because of the heavy use of ToStreamFactory so the additional memory load of the GetAllBytes call prior to the byte[] hash is near 0.  However, in cases where we do not have a MemoryStream and are working with large packages, we experience an unnecessary memory hit.
This large additional memory hit was seen in implementations that avoid the excessive MemoryStream creation
due to OOM errors on large packages. The stream method avoids this unnecessary memory load.

Comparing the stream and byte[] methods across each of the currently supported hash algorithms (times in seconds):

*For a large package (226M .nupkg) run 10 times
 [hashAlgo]                 [stream] [byte[]]
 SHA256CryptoServiceProvider 21.28    22.59
 SHA256Managed               38.17    39.07
 SHA512CryptoServiceProvider 14.01    14.77
 SHA512Managed               24.16    24.23

*For a small package (176K .nupkg) run 10000 times
 [hashAlgo]                 [stream] [byte[]]
 SHA256CryptoServiceProvider 16.46    17.26
 SHA256Managed               28.44    29.16
 SHA512CryptoServiceProvider 11.01    12.19
 SHA512Managed               17.99    19.82

For comparison purposes, both implementations used a preloaded MemoryStream as the data store, passing in the raw
stream for the stream methods, and using the NuGet StreamExtension ReadAllBytes() for the byte[] method.
Work Item: 2451