Parameterized existing command of View.PackageManagerConsole to invoke PSCmdlets from outside of VS Process
Added an interface IPowerConsoleService and implemented the interface in
PowerConsoleToolWindow. This has the logic to execute the PSCmdlets

Note that DTE.ExecuteCommand is an asynchronous call and does not
return success or failure. I share the sentiment of several developers
who feel this is a bad contract. But, here we are and I do the following

1. When the commmand is running, if execute is called, it will throw a
NotSupportedException("Power Console is busy at the moment")
2. Also, note that the Enabled value is toggled as
appropriate and the best way to check those values is to call QueryStatus.
This is the responsibility of the caller
3. Nevertheless, one can always send a full script
and not just 1 cmdlet as an input to a single call and wait for everything to finish.
This should do the trick for testing where not knowing when to run the
next command is the problem

Caveats
1. At the end, Always call a command 'clear' using the new parameter so that the
unflushed message in the PowerShell Console buffer is flushed out to VS
Output window
2. Using the new parameter to execute a command even before opening the
powershell console ONCE will issue an error. Rarely happens and ignore
that. Workaround by opening the console first