Fix of issue 2967.
The cause is that for a WiX project, its TargetFramework property is null.
This value is used as a key for a dictionary inside
VersionUtility.IsCompatible(), and this causes the observed error since
null is not allowed as key. The fix is to check for null at the beginning
of the method.