Fix of issue 2987 "Cannot install NuGet packages in WiX Projects 3.6/3.7"
The cause is that for some reason, project.UniqueName throws
COMException for Wix projects 3.6/3.7. BTW, this problem happens on
VS2012, but not on VS2010.

The fix is to use project.FullName when project.UniqueName throws
COMException.

The change also contains workaround for another problem with WiX project,
which is IsDocumentInProject() returns false even though the file is in
the WiX project.