Fix the problem that the pack command tests in class NuGetPackCommandTest cannot run successfully on CI build.
The cause is that on .NET 4.0, the
BuildManager.DefaultBuildManager.Build() method will set the working
directory to the working directory the last time when it's called.
Since that directory is deleted by the unit test, the
SetCurrentDirectory() call will fail, and the pack command will fail.

The fix is to execute the nuget.exe command instead of executng
Program.Main(). It seems we can easily get the location of the nuget.exe
by calling typeof(Program).Assembly.Location. However, xunit uses
ShadowCopy, and the location retuned contains file nuget.exe, but not
nuge.core.dll, which means this nuget.exe will not run because of missing
dependent dll error.

The solution I used is to create the app.config file before the tests are
run. The config file will contain a key called TargeDir whose value is the
target directory of the test project. Since all dependent DLLs are in the
target directory besides nuget.exe, unit tests can run that nuget.exe
successfully.