Fix of issue 3256: MSBuild import failures are not rolled back cleanly.
The problem is this:

- Because the import file is invalid, when the packages.config file is
added to the NativeProjectSystem, the add fails.

- During rollback, nuget needs to delete the packages.config
file. However, before the file is deleted, nuget first calls
RemoveFileFromProject() to try to remove the file from the project. Since
it's not added correctly in the first place, this step fails. In this
case, nuget will not delete the file, thus the problem we see.

The fix is to return success in RemoveFileFromProject() when
the file does not exist in the project, instead of return failure, this
current behavior. This way, nuget will continue to the next step and
delete the file from disk.

Note the tests show that add file fails only for NativeProjectSystem, so
there is no need to change the VsProjectSystem's implementation of DeleteFile().