Adds ability to parse so-called custom profile strings (like "net40-client") as well as expose the framework version that owns the profile (i.e. v4.0).
Portable profiles are parsed by iterating the folders in the reference assemblies path for .NETPortable, both for v4.0 and v4.5. But the target framework version (4.0 or 4.5) that a profile was loaded from is not exposed in the NetPortableProfile class.

This is useful for tools that leverage NuGet (or automatically generate packages) and need to compile tools, as this version is needed to set the TargetFrameworkVersion on an MSBuild project, for example.

Issue #2990