Fix memory leaks in UI.
Take class PackageRestoreBar as an example to illustrate the problem. In its
constructor, it's set up to listen to events of
PackageRestoreManager. However, there's no code to unlisten events. The
problem is that PackageRestoreManager is a long live object, thus every
time a PackageRestoreBar is created, it will be kept in memory forever
because it listens to PackageRestoreManager's events.

Also, when the PackageRestoreBar is created, the PackageManagerWindow is
set up to listen to PackageRestoreBar's events. Thus, this whole
reference chain keeps all PackageManagerWindows in memory.

Class ProductUpdateBar has similar problems.

The fixes:
- Add method CleanUp() to ProductUpdateBar & PackageRestoreBar to
unlisten events.
- In PackageManagerWindow.OnDialogWindowClosed() method, unlisten events
and call ProductUpdateBar.Cleanup() & PackageRestoreBar.Cleanup().
- Also, I need to remove the PackageRestoreBar & ProductUpdateBar from
PackageManagerWindow's LayoutRoot: WinDbg shows that when the
PackageRestoreBar is added to LayoutRoot, somehow, it will hold a
reference to PackageManagerWindow thru some DependencyProperty.