Build Warnings for Project Retargeting Added a bunch of unit tests and functional tests
Show Build warning when one or more packages in the project needs to be
reinstalled due to project retargeting. This is in addition to the
Build error shown ONCE when project retargeting takes place

Also, In order to increase the stability of the functional tests, I have
created a new package that targets net40 and net40-client

This way there are no additional warnings that show up when a project is
retargeting between net40 and net40-client. This used to happen when a
net40 project was retargeted to net35 making it hard to count the number
of warnings and retrieving the right warning item. With the new package,
only error or warning generated is related to project retargeting