Fix issues https://nuget.codeplex.com/workitem/3496, https://nuget.codeplex.com/workitem/3494
The problem arose because of a misunderstanding of the return value of VsUtility.GetFullPath(). The method returns the full path of the project directory, not the full path of the project file as I thought. The bug was discovered because the returned value happens to have a trailing slash for most project types. But for web site / JS metro project, there is no trailing slash in the return value, which exposes the bug.

Changes:
- Created new method GetPackageReferenceFileFullPath(project) that returns the correct full path of the packages.config file for the project.
- Fixes the same problem in project retargetting utility.
- Also contains fix for https://nuget.codeplex.com/workitem/3491: show message "All packages are already installed and there is nothing to restore." when neccessary.
- Added functional tests for package restore extension