Fix of issue https://nuget.codeplex.com/workitem/3568: opt-out message is displayed multiple times.
The cause: a boolean property is used so that the opt-out
message is displayed only once, but package restore can happen in
parallel and without a lock, it's still possible to display the message
multiple times.

Another change is that the package restore opt out message is displayed
only when -RequireCosent is specified.