Further fix of issue 3954.
The reason why "Install Antlr 3.5.0.2" is displayed twice is that the
first one is output by package manager to indicate installing the package
into the packages folder, while the second is output by
project manager to indicate installing the package into a project.

The fix is to change the output strings to make this clear.