Fix of issue 3678: problems in XDT causes the target file to be truncated.
The cause is that we use XmlTransformableDocument.Save(targetFile).
However, if an exception is thrown during save, the content of the targetFile can
get lost.

The fix is to save the xdt transform result into a memory stream first, then save
the memory stream into the target file. This way, if there's any problem with xdt,
the target file won't be touched.