Fix issue 3732: NuGet.exe install ignores "-NoCache" option.
The problem is not with nuget.exe, but with NuGet.Server. NuGet.Server caches
the hash of packages. However, if a package is pushed
to overwrite the existing package, the cached hash value is not deleted.
Since the hash doesn't change, nuget.exe thinks there's no difference between
the packages from NuGet.Server and the package in cache, so it continues to
use the package in cache.

The fix is to remove the cached metadata when a package is pushed.