Fix the performance problem in nuget.exe install.
The problem is that nuget.exe gets all versions of a package to find the
latest version, thus making many http calls, when it can just make one
call to get the latest version.

The cause of the problem:
- nuget.exe creates an AggregateRepository as the source repro;
- but AggregateRepository doesn't implement ILatestPackageLookup, so
  nuget falls back to the method of getting all packages to find the
  latest version.

The fix is to explicitly use ILatestPackageLookup on repositories inside
the AggregateRepository.