Fix of issue 4153: NuGet fails to include files
The behavior is that if the source file name contains "..", then this file will get
filtered out.  The problem is that when NuGet filters out package files, it should use the pattern
to match the file's path in the package. Instead, the current code tries to filter based
on the file's source path.