Adds support for specifying additional metadata via AssemblyMetadataAttribute
Currently the only way to augment the $replacement$ tokens is by passing additional values from the command line.
Just like for other assembly-level attributes, it's very convenient to be able to specify these additional key-value pairs using the AssemblyMetadataAttribute.
The command line arguments should always take precedence to those assembly attributes, so that behavior was added too.
Also, the package Id seemed an important property to be able to override also both as an AssemblyMetadataAttribute and a cmd line paramenter, so that was added too. Maybe it should be done for all manifest properties? See ProjectFactory.cs (214-236).