Fixed package sources loading from ISettings.
The problem is that <disabledPackageSources> specified in a config file
with lower priority should not affect sources specified in a config file
with higher priority. The fix is to add property Priority to class
SettingValue, and take this value into consideration when disabling package
sources.

Additionally, cleaned up ISettings interface.