Fix error pushing packages from Mono/Linux
MultipartWebRequest is using Environment.Newline in a couple places, which causes different payloads to be sent on Mono/Linux than on Windows. In our case, this was causing errors when pushing packages from Linux to a Sonatype Nexus artifact repository. This PR ensures that payload sent is the same (verified by using a packet sniffer) on Windows, Linux, and Mac.